+++ ihm/reader.py
@@ -1643,7 +1643,6 @@ class _ModelGroupHandler(Handler):
 
     def finalize(self):
         # Put all model groups not assigned to a state in their own state
-        # todo: handle models not in model groups too?
         model_groups_in_states = set()
         for sg in self.system.state_groups:
             for state in sg:
@@ -1653,6 +1652,19 @@ class _ModelGroupHandler(Handler):
                if mgid not in model_groups_in_states]
         if mgs:
             s = ihm.model.State(mgs)
+            self.system.state_groups.append(ihm.model.StateGroup([s]))
+
+        # Put all models not in a group in their own group in its own state
+        # (e.g. this will catch models from a non-IHM file)
+        models_in_groups = set()
+        for mg in self.sysr.model_groups._obj_by_id.values():
+            for m in mg:
+                models_in_groups.add(m._id)
+        ms = [m for mid, m in self.sysr.models._obj_by_id.items()
+              if mid not in models_in_groups]
+        if ms:
+            mg = ihm.model.ModelGroup(ms)
+            s = ihm.model.State([mg])
             self.system.state_groups.append(ihm.model.StateGroup([s]))